@@ -1,34 +1,70 @@
 {
-  "name": "column-controls",
-  "version": "1.0.0",
-  "description": "table column formaters",
+  "name": "signature-pad",
+  "version": "2.0.0",
+  "description": "signature-pad",
   "main": "index.js",
   "scripts": {
-    "dev": "webpack-dev-server",
-    "build": "webpack -p",
+    "start": "webpack-dev-server",
+    "build": "webpack -p --config webpack.production.config.js",
+    "build-no-source-map": "webpack -p --config webpack.production.config.js --devtool none",
+    "upload": "node upload.js",
+    "test-unit": "jest --watch",
+    "test-integration": "testcafe-live chrome ./src/__tests__/integration",
+    "test-integration-remote": "testcafe-live remote ./src/__tests__/integration --qr-code",
+    "update": "node update.js",
     "debug": "ngrok http --host-header=rewrite 8080"
   },
   "author": "Dell Boomi",
   "license": "MIT",
   "dependencies": {
-    "@types/react": "^16.0.38",
-    "@types/react-dom": "^16.0.4",
-    "moment": "^2.22.2",
-    "npm": "^6.4.0",
-    "prop-types": "^15.6.2",
-    "react": "^16.2.0",
-    "react-dom": "^16.2.0",
-    "react-signature-pad-wrapper": "^1.2.2"
+    "@types/jquery": "^3.3.29",
+    "acorn": "^6.1.1",
+    "dedent": "^0.7.0",
+    "react": "^16.3.2",
+    "react-dom": "^16.3.2"
   },
   "devDependencies": {
-    "awesome-typescript-loader": "^3.4.1",
-    "css-loader": "^0.28.9",
-    "extract-text-webpack-plugin": "^3.0.2",
+    "@types/jest": "^23.1.4",
+    "@types/loglevel": "^1.5.3",
+    "@types/react": "^16.7.3",
+    "@types/react-dom": "^16.0.9",
+    "awesome-typescript-loader": "5.2.0",
+    "clipboardy": "^1.2.3",
+    "colors": "^1.3.1",
+    "css-loader": "^2.1.1",
+    "download": "^7.1.0",
+    "fs-extra": "^7.0.1",
+    "glob-promise": "^3.4.0",
+    "inquirer": "^6.0.0",
+    "inquirer-autocomplete-prompt": "^1.0.1",
+    "mime-types": "^2.1.19",
+    "mini-css-extract-plugin": "^0.5.0",
+    "react-test-renderer": "^16.4.1",
+    "recursive-copy": "^2.0.9",
+    "request-promise-native": "^1.0.5",
     "source-map-loader": "^0.2.3",
-    "style-loader": "^0.20.2",
-    "typescript": "^2.7.2",
-    "webpack": "^3.11.0",
-    "webpack-dev-server": "^2.11.1",
-    "write-file-webpack-plugin": "^4.2.0"
-  }
+    "style-loader": "0.21.0",
+    "three-way-merger": "^0.5.0",
+    "tslint": "^5.10.0",
+    "tslint-loader": "3.6.0",
+    "tslint-react": "^3.6.0",
+    "typescript": "^3.4.2",
+    "webpack": "^4.27.1",
+    "webpack-cli": "^3.1.2",
+    "webpack-dev-server": "^3.1.14",
+    "write-file-webpack-plugin": "4.4.1",
+    "react-signature-pad-wrapper": "^1.2.2"
+  },
+  "flow": {
+    "tenantId": "d1b3e73f-3f2a-4159-8fbb-ceb11a84e28d",
+    "id": "88b7b053-049b-402d-b614-597bed6a4b78",
+    "versionId": "",
+    "filenames": {
+      "js": "SignatureBox.js",
+      "css": "SignatureBox.css"
+    }
+  },
+  "peerDependencies": {},
+  "bundledDependencies": [],
+  "optionalDependencies": {}
 }
