@@ -1,130 +1,63 @@
-
-declare var manywho: any;
-
 import * as React from 'react';
-var SignaturePad  = require('react-signature-pad-wrapper');
-import './SignatureBox.css';
+import SignaturePad  from 'react-signature-pad-wrapper';
+import './css/SignatureBox.css';
+import { FlowComponent } from '/Operational Data/Flow UI Custom Components/2019 Version/FlowComponentModel/src/FlowComponent';
+import { IManywho } from '/Operational Data/Flow UI Custom Components/2019 Version/FlowComponentModel/src/interfaces';
 
-class SignatureBox extends React.Component<any, any> 
-{   
-    componentId: string = "";
-    flowKey: string ="";    
-    attributes : any = {};
-    selectedItem: string = null;
+declare const manywho: IManywho;
 
-    text : string = "";
+class SignatureBox extends FlowComponent {
 
-    constructor(props : any)
-	{
+    sigpad: any;
+
+    constructor(props: any) {
         super(props);
-        
-        this.componentId = props.id;
-        this.flowKey = props.flowKey;
 
-        //push attributes into keyed map 
-		var flowModel = manywho.model.getComponent(this.props.id,   this.props.flowKey);
-		if(flowModel.attributes)
-		{
-			for(var key in flowModel.attributes)
-			{
-				this.attributes[key] = flowModel.attributes[key];
-			}
-        }
+        this.clearSignature = this.clearSignature.bind(this);
+        this.strokeEnd = this.strokeEnd.bind(this);
     }
 
-    
-    componentDidMount() 
-    {
-        var flowModel = manywho.model.getComponent(this.props.id,   this.props.flowKey);
-        const flowState = manywho.state.getComponent(this.props.id,   this.props.flowKey);
-
-        if(flowModel.contentValue)
-		{
-            flowState.contentValue = flowModel.contentValue;
-            var sigPad : any = this.refs.signaturePad;
-            sigPad.fromDataURL(flowModel.contentValue);
-        }
-
-        this.forceUpdate();
+    async componentDidMount() {
+        await super.componentDidMount();
     }
 
-    componentDidUpdate()
-    {
-
-        var sigPad : any = this.refs.signaturePad;
-        //sigPad.onEnd = this.strokeEnd.bind(this); 
+    clearSignature() {
+        this.sigpad.clear();
+        const data: string = this.sigpad.toDataURL();
+        this.setStateValue(data);
     }
 
-	getAttribute(attributeName : string)
-	{
-		if(this.attributes[attributeName])
-		{
-			return this.attributes[attributeName];
-		}
-		else
-		{
-			return null;
-		}
-	}
-
-       
-    clearSignature()
-    {
-        var sigpad : any = this.refs.signaturePad;
-        sigpad.clear();
-
-        var data : string = sigpad.toDataURL();
-        
-        const flowState = manywho.state.getComponent(this.props.id,   this.props.flowKey);
-
-        flowState.contentValue = data;
-
+    strokeEnd() {
+        const data: string = this.sigpad.toDataURL();
+        this.setStateValue(data);
     }
 
-    strokeEnd()
-    {
-        var sigpad : any = this.refs.signaturePad;
+	   render() {
+        const options: any = {};
+        options.width = '100px';
+        options.height = '70';
+        options.penColor = '#000000';
+        options.redrawOnResize = true;
+        options.onEnd = this.strokeEnd;
 
-        var data : string = sigpad.toDataURL();
-        
-        const flowState = manywho.state.getComponent(this.props.id,   this.props.flowKey);
+        const title = this.getAttribute('title') || 'Signature - please draw your signature in the area below';
 
-        flowState.contentValue = data;
-    }
-
-
-	render()
-	{
-	   
-		const flowModel = manywho.model.getComponent(this.props.id,   this.props.flowKey);
-        const flowState = manywho.state.getComponent(this.props.id,   this.props.flowKey);
-        
-        var options : any = {};
-        options.width="100px";
-        options.height="70";
-        options.penColor = "#000000";
-        options.redrawOnResize= true;
-        options.onEnd = this.strokeEnd.bind(this);
-    
-        var self = this;
-
-        var title = this.getAttribute("Title") || "Signature - please draw your signature in the area below";
-        
-
-        return <div className="signature-box">
-                    <div className="signature-box-header">
-                        <div style={{float:'left'}}>
-                            <span className="signature-box-header-title">{title}}</span>
-                        </div>
-                        <div style={{float:'right'}}>
-                            <span className="glyphicon glyphicon-remove signature-box-header-button" onClick={this.clearSignature.bind(this)}></span>
-                        </div>
-                        
+        return (
+            <div className="signature-box">
+                <div className="signature-box-header">
+                    <div style={{float: 'left'}}>
+                        <span className="signature-box-header-title">{title}</span>
                     </div>
-                    <div className="signature-box-body">
-                        <SignaturePad ref="signaturePad" redrawOnResize={true} options={options} />
+                    <div style={{float: 'right'}}>
+                        <span className="glyphicon glyphicon-remove signature-box-header-button" onClick={this.clearSignature}/>
                     </div>
-               </div>
+
+                </div>
+                <div className="signature-box-body">
+                    <SignaturePad ref={(ele: any) => {this.sigpad = ele; }} redrawOnResize={true} options={options}/>
+                </div>
+            </div>
+        );
     }
 }
 
